SG_code/requirements.txt
audioread==3.0.1
easydict==1.13
einops==0.8.0
h5py==3.11.0
joblib==1.4.2
json5==0.9.6
librosa==0.10.2.post1
matplotlib==3.9.2
numpy==1.26.4
pillow==10.4.0
tensorboardx==2.6.2.2
tqdm==4.66.5
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.1
soundfile==0.12.1
openai==1.31.0
transformers==4.47.1
vllm==0.6.6.post1